small validation fixes
