+++ src/Components/Users/UsersForm.js
@@ -6,10 +6,10 @@ import {
   nameValidationFourLength, 
   passwordValidationEightLength, 
   fileValidationExtensions
-} from "../../Services/formValidationService";
+} from "../../Services/formValidationsService";
 import {
-  postUser,
-  putUser
+  updateUserData, 
+  createUser
 } from "../../Services/userApiService";
 
 const UserForm = ({prevUserData}) => {
@@ -43,11 +43,23 @@ const UserForm = ({prevUserData}) => {
   const fileHandlerValidation = async (e) => {
     const file = e.target.files[0];
     const errorMessage = fileValidationExtensions(e);
-    if(errorMessage){
-      setFileError(errorMessage);
-    } 
-    if(errorMessage === undefined) {
+    if(!errorMessage){
       setFile(window.btoa(encodeURIComponent( file )));
+    } else {
+      setFileError(errorMessage);
+    }
+  };
+
+  const onSubmitHandler = (values) => {
+    const newUserData = {
+      ...values, 
+      role_id: parseInt(values.role_id), 
+      profile_image : file
+    };
+    if(prevUserDataExist){
+      updateUserData(newUserData);
+    } else {
+      createUser(newUserData);
     }
   };
 
@@ -55,68 +67,61 @@ const UserForm = ({prevUserData}) => {
     <Formik
      initialValues={ prevUserData || initialUserData }
      validate= {validate}
-     onSubmit={(values) => {
-       const newUserData = {...values, role_id: parseInt(values.role_id), profile_image : file};
-      if(prevUserDataExist){
-        putUser(newUserData);
-      } else {
-        postUser(newUserData);
-      }
-     }}
-   >
-     {({
+     onSubmit={onSubmitHandler}
+    >
+      {({
        values,
        errors,
        handleChange,
        handleSubmit
-     }) => (
+      }) => (
       <form className="form-container" onSubmit={handleSubmit}>
-          <Field
-            className="input-field" 
-            name="name" 
-            onChange={handleChange}
-            value={values.name}
-            placeholder="Nombre"
-          />
-          {errors.name && errors.name}
-          <Field
-            className="input-field" 
-            name="email" 
-            onChange={handleChange}
-            value={values.email}
-            placeholder="Email"
-          />
-          {errors.email && errors.email}
-          <input 
+        <Field
+          className="input-field" 
+          name="name" 
+          onChange={handleChange}
+          value={values.name}
+          placeholder="Nombre"
+        />
+        {errors.name && errors.name}
+        <Field
+          className="input-field" 
+          name="email" 
+          onChange={handleChange}
+          value={values.email}
+          placeholder="Email"
+        />
+        {errors.email && errors.email}
+        <input 
           required
           name="file"
           type="file"
           className="input-field"
           onChange={fileHandlerValidation}
-          />
-          {fileError && <p>{fileError}</p>}
-          <Field 
-            required
-            name="role_id"
-            as="select"
-            className="input-field" 
-            value={values.role_id} 
-            onChange={handleChange}>
-              <option value="" disabled >Escoja su rol</option>
-              <option value="1">Admin</option>
-              <option value="2">User</option>
-          </Field>
-          <Field
-            className="input-field" 
-            type="password"
-            name="password"
-            placeholder='Contraseña'
-            onChange={handleChange}
-            value={values.password}
-          />
-          {errors.password && errors.password}
-          <button className="submit-btn" type="submit">Send</button>
-        </form>
+        />
+        {fileError && <p>{fileError}</p>}
+        <Field 
+          required
+          name="role_id"
+          as="select"
+          className="input-field" 
+          value={values.role_id} 
+          onChange={handleChange}>
+            <option value="" disabled >Escoja su rol</option>
+            <option value="1">Admin</option>
+            <option value="2">User</option>
+        </Field>
+        <Field
+          className="input-field" 
+          type="password"
+          name="password"
+          placeholder='Contraseña'
+          onChange={handleChange}
+          value={values.password}
+        />
+        {errors.password && errors.password}
+        <button className="submit-btn" type="submit">Send</button>
+      </form>
      )}
     </Formik>
   );
